requirements.txt
nltk==3.2.1
numpy==1.10.1
scipy>=0.9
scikit-learn==0.17.1
Django==1.9
python-social-auth>=0.2.19
djangorestframework>=3.3.3